Move commands to separated modules
