+++ framework/lib/commands/Modules/index.ts
@@ -1,0 +1,5 @@
+export * from "./AboutCommand";
+export * from "./ConfigCommand";
+export * from "./PingCommand";
+export * from "./ReadMangaCommand";
+export * from "./SearchMangaCommand";
